separating code that will be reused
