+++ src/vs/editor/test/browser/controller/imeTester.ts
@@ -3,8 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { ITextAreaInputHost, TextAreaInput, TextAreaWrapper } from 'vs/editor/browser/controller/editContext/textArea/textAreaInput';
-import { ISimpleModel, PagedScreenReaderStrategy, TextAreaState } from 'vs/editor/browser/controller/editContext/textArea/textAreaState';
+import { TextAreaWrapper } from 'vs/editor/browser/controller/editContext/textArea/textAreaInput';
 import { Position } from 'vs/editor/common/core/position';
 import { IRange, Range } from 'vs/editor/common/core/range';
 import { EndOfLinePreference } from 'vs/editor/common/model';
@@ -14,6 +13,8 @@ import * as platform from 'vs/base/common/platform';
 import { mainWindow } from 'vs/base/browser/window';
 import { TestAccessibilityService } from 'vs/platform/accessibility/test/common/testAccessibilityService';
 import { NullLogService } from 'vs/platform/log/common/log';
+import { HiddenAreaState, ISimpleModel, PagedScreenReaderStrategy } from 'vs/editor/browser/controller/editContext/editContextState';
+import { HiddenAreaInput, IHiddenAreaInputHost } from 'vs/editor/browser/controller/editContext/editContextInput';
 
 // To run this test, open imeTester.html
 
@@ -104,7 +105,7 @@ function doCreateTest(description: string, inputStr: string, expectedStr: string
 
 	const model = new SingleLineTestModel('some  text');
 
-	const textAreaInputHost: ITextAreaInputHost = {
+	const textAreaInputHost: IHiddenAreaInputHost = {
 		getDataToCopy: () => {
 			return {
 				isFromEmptySelection: false,
@@ -114,7 +115,7 @@ function doCreateTest(description: string, inputStr: string, expectedStr: string
 				mode: null
 			};
 		},
-		getScreenReaderContent: (): TextAreaState => {
+		getScreenReaderContent: (): HiddenAreaState => {
 			const selection = new Range(1, 1 + cursorOffset, 1, 1 + cursorOffset + cursorLength);
 
 			return PagedScreenReaderStrategy.fromEditorSelection(model, selection, 10, true);
@@ -124,7 +125,7 @@ function doCreateTest(description: string, inputStr: string, expectedStr: string
 		}
 	};
 
-	const handler = new TextAreaInput(textAreaInputHost, new TextAreaWrapper(input), platform.OS, {
+	const handler = new HiddenAreaInput(textAreaInputHost, new TextAreaWrapper(input), platform.OS, {
 		isAndroid: browser.isAndroid,
 		isFirefox: browser.isFirefox,
 		isChrome: browser.isChrome,
