+++ src/vs/editor/test/browser/controller/textAreaState.test.ts
@@ -6,12 +6,12 @@
 import assert from 'assert';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { ensureNoDisposablesAreLeakedInTestSuite } from 'vs/base/test/common/utils';
-import { ITextAreaWrapper, PagedScreenReaderStrategy, TextAreaState } from 'vs/editor/browser/controller/editContext/textArea/textAreaState';
+import { HiddenAreaState, IHiddenAreaWrapper, PagedScreenReaderStrategy } from 'vs/editor/browser/controller/editContext/editContextState';
 import { Range } from 'vs/editor/common/core/range';
 import { Selection } from 'vs/editor/common/core/selection';
 import { createTextModel } from 'vs/editor/test/common/testTextModel';
 
-class MockTextAreaWrapper extends Disposable implements ITextAreaWrapper {
+class MockTextAreaWrapper extends Disposable implements IHiddenAreaWrapper {
 
 	public _value: string;
 	public _selectionStart: number;
@@ -60,7 +60,7 @@ class MockTextAreaWrapper extends Disposable implements ITextAreaWrapper {
 	}
 }
 
-function equalsTextAreaState(a: TextAreaState, b: TextAreaState): boolean {
+function equalsTextAreaState(a: HiddenAreaState, b: HiddenAreaState): boolean {
 	return (
 		a.value === b.value
 		&& a.selectionStart === b.selectionStart
@@ -74,8 +74,8 @@ suite('TextAreaState', () => {
 
 	ensureNoDisposablesAreLeakedInTestSuite();
 
-	function assertTextAreaState(actual: TextAreaState, value: string, selectionStart: number, selectionEnd: number): void {
-		const desired = new TextAreaState(value, selectionStart, selectionEnd, null, undefined);
+	function assertTextAreaState(actual: HiddenAreaState, value: string, selectionStart: number, selectionEnd: number): void {
+		const desired = new HiddenAreaState(value, selectionStart, selectionEnd, null, undefined);
 		assert.ok(equalsTextAreaState(desired, actual), desired.toString() + ' == ' + actual.toString());
 	}
 
@@ -84,7 +84,7 @@ suite('TextAreaState', () => {
 		textArea._value = 'Hello world!';
 		textArea._selectionStart = 1;
 		textArea._selectionEnd = 12;
-		let actual = TextAreaState.readFromTextArea(textArea, null);
+		let actual = HiddenAreaState.readFromTextArea(textArea, null);
 
 		assertTextAreaState(actual, 'Hello world!', 1, 12);
 		assert.strictEqual(actual.value, 'Hello world!');
@@ -102,21 +102,21 @@ suite('TextAreaState', () => {
 		textArea._selectionStart = 1;
 		textArea._selectionEnd = 12;
 
-		let state = new TextAreaState('Hi world!', 2, 2, null, undefined);
+		let state = new HiddenAreaState('Hi world!', 2, 2, null, undefined);
 		state.writeToTextArea('test', textArea, false);
 
 		assert.strictEqual(textArea._value, 'Hi world!');
 		assert.strictEqual(textArea._selectionStart, 9);
 		assert.strictEqual(textArea._selectionEnd, 9);
 
-		state = new TextAreaState('Hi world!', 3, 3, null, undefined);
+		state = new HiddenAreaState('Hi world!', 3, 3, null, undefined);
 		state.writeToTextArea('test', textArea, false);
 
 		assert.strictEqual(textArea._value, 'Hi world!');
 		assert.strictEqual(textArea._selectionStart, 9);
 		assert.strictEqual(textArea._selectionEnd, 9);
 
-		state = new TextAreaState('Hi world!', 0, 2, null, undefined);
+		state = new HiddenAreaState('Hi world!', 0, 2, null, undefined);
 		state.writeToTextArea('test', textArea, true);
 
 		assert.strictEqual(textArea._value, 'Hi world!');
@@ -126,16 +126,16 @@ suite('TextAreaState', () => {
 		textArea.dispose();
 	});
 
-	function testDeduceInput(prevState: TextAreaState | null, value: string, selectionStart: number, selectionEnd: number, couldBeEmojiInput: boolean, expected: string, expectedCharReplaceCnt: number): void {
-		prevState = prevState || TextAreaState.EMPTY;
+	function testDeduceInput(prevState: HiddenAreaState | null, value: string, selectionStart: number, selectionEnd: number, couldBeEmojiInput: boolean, expected: string, expectedCharReplaceCnt: number): void {
+		prevState = prevState || HiddenAreaState.EMPTY;
 
 		const textArea = new MockTextAreaWrapper();
 		textArea._value = value;
 		textArea._selectionStart = selectionStart;
 		textArea._selectionEnd = selectionEnd;
 
-		const newState = TextAreaState.readFromTextArea(textArea, null);
-		const actual = TextAreaState.deduceInput(prevState, newState, couldBeEmojiInput);
+		const newState = HiddenAreaState.readFromTextArea(textArea, null);
+		const actual = HiddenAreaState.deduceInput(prevState, newState, couldBeEmojiInput);
 
 		assert.deepStrictEqual(actual, {
 			text: expected,
@@ -158,7 +158,7 @@ suite('TextAreaState', () => {
 
 	test('issue #2586: Replacing selected end-of-line with newline locks up the document', () => {
 		testDeduceInput(
-			new TextAreaState(']\n', 1, 2, null, undefined),
+			new HiddenAreaState(']\n', 1, 2, null, undefined),
 			']\n',
 			2, 2, true,
 			'\n', 0
@@ -176,7 +176,7 @@ suite('TextAreaState', () => {
 
 	test('extractNewText - typing does not cause a selection', () => {
 		testDeduceInput(
-			TextAreaState.EMPTY,
+			HiddenAreaState.EMPTY,
 			'a',
 			0, 1, true,
 			'a', 0
@@ -185,7 +185,7 @@ suite('TextAreaState', () => {
 
 	test('extractNewText - had the textarea empty', () => {
 		testDeduceInput(
-			TextAreaState.EMPTY,
+			HiddenAreaState.EMPTY,
 			'a',
 			1, 1, true,
 			'a', 0
@@ -194,7 +194,7 @@ suite('TextAreaState', () => {
 
 	test('extractNewText - had the entire line selected', () => {
 		testDeduceInput(
-			new TextAreaState('Hello world!', 0, 12, null, undefined),
+			new HiddenAreaState('Hello world!', 0, 12, null, undefined),
 			'H',
 			1, 1, true,
 			'H', 0
@@ -203,7 +203,7 @@ suite('TextAreaState', () => {
 
 	test('extractNewText - had previous text 1', () => {
 		testDeduceInput(
-			new TextAreaState('Hello world!', 12, 12, null, undefined),
+			new HiddenAreaState('Hello world!', 12, 12, null, undefined),
 			'Hello world!a',
 			13, 13, true,
 			'a', 0
@@ -212,7 +212,7 @@ suite('TextAreaState', () => {
 
 	test('extractNewText - had previous text 2', () => {
 		testDeduceInput(
-			new TextAreaState('Hello world!', 0, 0, null, undefined),
+			new HiddenAreaState('Hello world!', 0, 0, null, undefined),
 			'aHello world!',
 			1, 1, true,
 			'a', 0
@@ -221,7 +221,7 @@ suite('TextAreaState', () => {
 
 	test('extractNewText - had previous text 3', () => {
 		testDeduceInput(
-			new TextAreaState('Hello world!', 6, 11, null, undefined),
+			new HiddenAreaState('Hello world!', 6, 11, null, undefined),
 			'Hello other!',
 			11, 11, true,
 			'other', 0
@@ -230,7 +230,7 @@ suite('TextAreaState', () => {
 
 	test('extractNewText - IME', () => {
 		testDeduceInput(
-			TextAreaState.EMPTY,
+			HiddenAreaState.EMPTY,
 			'これは',
 			3, 3, true,
 			'これは', 0
@@ -239,7 +239,7 @@ suite('TextAreaState', () => {
 
 	test('extractNewText - isInOverwriteMode', () => {
 		testDeduceInput(
-			new TextAreaState('Hello world!', 0, 0, null, undefined),
+			new HiddenAreaState('Hello world!', 0, 0, null, undefined),
 			'Aello world!',
 			1, 1, true,
 			'A', 0
@@ -248,7 +248,7 @@ suite('TextAreaState', () => {
 
 	test('extractMacReplacedText - does nothing if there is selection', () => {
 		testDeduceInput(
-			new TextAreaState('Hello world!', 5, 5, null, undefined),
+			new HiddenAreaState('Hello world!', 5, 5, null, undefined),
 			'Hellö world!',
 			4, 5, true,
 			'ö', 0
@@ -257,7 +257,7 @@ suite('TextAreaState', () => {
 
 	test('extractMacReplacedText - does nothing if there is more than one extra char', () => {
 		testDeduceInput(
-			new TextAreaState('Hello world!', 5, 5, null, undefined),
+			new HiddenAreaState('Hello world!', 5, 5, null, undefined),
 			'Hellöö world!',
 			5, 5, true,
 			'öö', 1
@@ -266,7 +266,7 @@ suite('TextAreaState', () => {
 
 	test('extractMacReplacedText - does nothing if there is more than one changed char', () => {
 		testDeduceInput(
-			new TextAreaState('Hello world!', 5, 5, null, undefined),
+			new HiddenAreaState('Hello world!', 5, 5, null, undefined),
 			'Helöö world!',
 			5, 5, true,
 			'öö', 2
@@ -275,7 +275,7 @@ suite('TextAreaState', () => {
 
 	test('extractMacReplacedText', () => {
 		testDeduceInput(
-			new TextAreaState('Hello world!', 5, 5, null, undefined),
+			new HiddenAreaState('Hello world!', 5, 5, null, undefined),
 			'Hellö world!',
 			5, 5, true,
 			'ö', 1
@@ -284,7 +284,7 @@ suite('TextAreaState', () => {
 
 	test('issue #25101 - First key press ignored', () => {
 		testDeduceInput(
-			new TextAreaState('a', 0, 1, null, undefined),
+			new HiddenAreaState('a', 0, 1, null, undefined),
 			'a',
 			1, 1, true,
 			'a', 0
@@ -293,7 +293,7 @@ suite('TextAreaState', () => {
 
 	test('issue #16520 - Cmd-d of single character followed by typing same character as has no effect', () => {
 		testDeduceInput(
-			new TextAreaState('x x', 0, 1, null, undefined),
+			new HiddenAreaState('x x', 0, 1, null, undefined),
 			'x x',
 			1, 1, true,
 			'x', 0
@@ -301,18 +301,18 @@ suite('TextAreaState', () => {
 	});
 
 	function testDeduceAndroidCompositionInput(
-		prevState: TextAreaState | null,
+		prevState: HiddenAreaState | null,
 		value: string, selectionStart: number, selectionEnd: number,
 		expected: string, expectedReplacePrevCharCnt: number, expectedReplaceNextCharCnt: number, expectedPositionDelta: number): void {
-		prevState = prevState || TextAreaState.EMPTY;
+		prevState = prevState || HiddenAreaState.EMPTY;
 
 		const textArea = new MockTextAreaWrapper();
 		textArea._value = value;
 		textArea._selectionStart = selectionStart;
 		textArea._selectionEnd = selectionEnd;
 
-		const newState = TextAreaState.readFromTextArea(textArea, null);
-		const actual = TextAreaState.deduceAndroidCompositionInput(prevState, newState);
+		const newState = HiddenAreaState.readFromTextArea(textArea, null);
+		const actual = HiddenAreaState.deduceAndroidCompositionInput(prevState, newState);
 
 		assert.deepStrictEqual(actual, {
 			text: expected,
@@ -326,7 +326,7 @@ suite('TextAreaState', () => {
 
 	test('Android composition input 1', () => {
 		testDeduceAndroidCompositionInput(
-			new TextAreaState('Microsoft', 4, 4, null, undefined),
+			new HiddenAreaState('Microsoft', 4, 4, null, undefined),
 			'Microsoft',
 			4, 4,
 			'', 0, 0, 0,
@@ -335,7 +335,7 @@ suite('TextAreaState', () => {
 
 	test('Android composition input 2', () => {
 		testDeduceAndroidCompositionInput(
-			new TextAreaState('Microsoft', 4, 4, null, undefined),
+			new HiddenAreaState('Microsoft', 4, 4, null, undefined),
 			'Microsoft',
 			0, 9,
 			'', 0, 0, 5,
@@ -344,7 +344,7 @@ suite('TextAreaState', () => {
 
 	test('Android composition input 3', () => {
 		testDeduceAndroidCompositionInput(
-			new TextAreaState('Microsoft', 0, 9, null, undefined),
+			new HiddenAreaState('Microsoft', 0, 9, null, undefined),
 			'Microsoft\'s',
 			11, 11,
 			'\'s', 0, 0, 0,
@@ -353,7 +353,7 @@ suite('TextAreaState', () => {
 
 	test('Android backspace', () => {
 		testDeduceAndroidCompositionInput(
-			new TextAreaState('undefinedVariable', 2, 2, null, undefined),
+			new HiddenAreaState('undefinedVariable', 2, 2, null, undefined),
 			'udefinedVariable',
 			1, 1,
 			'', 1, 0, 0,
@@ -362,7 +362,7 @@ suite('TextAreaState', () => {
 
 	suite('PagedScreenReaderStrategy', () => {
 
-		function testPagedScreenReaderStrategy(lines: string[], selection: Selection, expected: TextAreaState): void {
+		function testPagedScreenReaderStrategy(lines: string[], selection: Selection, expected: HiddenAreaState): void {
 			const model = createTextModel(lines.join('\n'));
 			const actual = PagedScreenReaderStrategy.fromEditorSelection(model, selection, 10, true);
 			assert.ok(equalsTextAreaState(actual, expected));
@@ -375,7 +375,7 @@ suite('TextAreaState', () => {
 					'Hello world!'
 				],
 				new Selection(1, 13, 1, 13),
-				new TextAreaState('Hello world!', 12, 12, new Range(1, 13, 1, 13), 0)
+				new HiddenAreaState('Hello world!', 12, 12, new Range(1, 13, 1, 13), 0)
 			);
 
 			testPagedScreenReaderStrategy(
@@ -383,7 +383,7 @@ suite('TextAreaState', () => {
 					'Hello world!'
 				],
 				new Selection(1, 1, 1, 1),
-				new TextAreaState('Hello world!', 0, 0, new Range(1, 1, 1, 1), 0)
+				new HiddenAreaState('Hello world!', 0, 0, new Range(1, 1, 1, 1), 0)
 			);
 
 			testPagedScreenReaderStrategy(
@@ -391,7 +391,7 @@ suite('TextAreaState', () => {
 					'Hello world!'
 				],
 				new Selection(1, 1, 1, 6),
-				new TextAreaState('Hello world!', 0, 5, new Range(1, 1, 1, 6), 0)
+				new HiddenAreaState('Hello world!', 0, 5, new Range(1, 1, 1, 6), 0)
 			);
 		});
 
@@ -402,7 +402,7 @@ suite('TextAreaState', () => {
 					'How are you?'
 				],
 				new Selection(1, 1, 1, 1),
-				new TextAreaState('Hello world!\nHow are you?', 0, 0, new Range(1, 1, 1, 1), 0)
+				new HiddenAreaState('Hello world!\nHow are you?', 0, 0, new Range(1, 1, 1, 1), 0)
 			);
 
 			testPagedScreenReaderStrategy(
@@ -411,7 +411,7 @@ suite('TextAreaState', () => {
 					'How are you?'
 				],
 				new Selection(2, 1, 2, 1),
-				new TextAreaState('Hello world!\nHow are you?', 13, 13, new Range(2, 1, 2, 1), 1)
+				new HiddenAreaState('Hello world!\nHow are you?', 13, 13, new Range(2, 1, 2, 1), 1)
 			);
 		});
 
@@ -421,7 +421,7 @@ suite('TextAreaState', () => {
 					'L1\nL2\nL3\nL4\nL5\nL6\nL7\nL8\nL9\nL10\nL11\nL12\nL13\nL14\nL15\nL16\nL17\nL18\nL19\nL20\nL21'
 				],
 				new Selection(1, 1, 1, 1),
-				new TextAreaState('L1\nL2\nL3\nL4\nL5\nL6\nL7\nL8\nL9\nL10\n', 0, 0, new Range(1, 1, 1, 1), 0)
+				new HiddenAreaState('L1\nL2\nL3\nL4\nL5\nL6\nL7\nL8\nL9\nL10\n', 0, 0, new Range(1, 1, 1, 1), 0)
 			);
 
 			testPagedScreenReaderStrategy(
@@ -429,7 +429,7 @@ suite('TextAreaState', () => {
 					'L1\nL2\nL3\nL4\nL5\nL6\nL7\nL8\nL9\nL10\nL11\nL12\nL13\nL14\nL15\nL16\nL17\nL18\nL19\nL20\nL21'
 				],
 				new Selection(11, 1, 11, 1),
-				new TextAreaState('L11\nL12\nL13\nL14\nL15\nL16\nL17\nL18\nL19\nL20\n', 0, 0, new Range(11, 1, 11, 1), 0)
+				new HiddenAreaState('L11\nL12\nL13\nL14\nL15\nL16\nL17\nL18\nL19\nL20\n', 0, 0, new Range(11, 1, 11, 1), 0)
 			);
 
 			testPagedScreenReaderStrategy(
@@ -437,7 +437,7 @@ suite('TextAreaState', () => {
 					'L1\nL2\nL3\nL4\nL5\nL6\nL7\nL8\nL9\nL10\nL11\nL12\nL13\nL14\nL15\nL16\nL17\nL18\nL19\nL20\nL21'
 				],
 				new Selection(12, 1, 12, 1),
-				new TextAreaState('L11\nL12\nL13\nL14\nL15\nL16\nL17\nL18\nL19\nL20\n', 4, 4, new Range(12, 1, 12, 1), 1)
+				new HiddenAreaState('L11\nL12\nL13\nL14\nL15\nL16\nL17\nL18\nL19\nL20\n', 4, 4, new Range(12, 1, 12, 1), 1)
 			);
 
 			testPagedScreenReaderStrategy(
@@ -445,7 +445,7 @@ suite('TextAreaState', () => {
 					'L1\nL2\nL3\nL4\nL5\nL6\nL7\nL8\nL9\nL10\nL11\nL12\nL13\nL14\nL15\nL16\nL17\nL18\nL19\nL20\nL21'
 				],
 				new Selection(21, 1, 21, 1),
-				new TextAreaState('L21', 0, 0, new Range(21, 1, 21, 1), 0)
+				new HiddenAreaState('L21', 0, 0, new Range(21, 1, 21, 1), 0)
 			);
 		});
